server/requirements.txt
fastapi
uvicorn
aiortc
stem 
dotenv
pynacl
pytest
httpx
requests
starlette
requests[socks]
websockets
aiohttp
aiohttp-socks